# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "ApplicationListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ApplicationResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ApplicationListResponse",
        "type": "object"
      },
      "ApplicationResponse": {
        "properties": {
          "apply_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apply Url"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "cover_letter_draft": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Letter Draft"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_remote": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Remote"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "notes": {
            "title": "Notes",
            "type": "string"
          },
          "opportunity_id": {
            "format": "uuid",
            "title": "Opportunity Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "opportunity_id",
          "status",
          "notes",
          "cover_letter_draft",
          "created_at",
          "updated_at"
        ],
        "title": "ApplicationResponse",
        "type": "object"
      },
      "ApplicationStatus": {
        "enum": [
          "interested",
          "prepared",
          "applied",
          "interviewing",
          "offer",
          "rejected",
          "withdrawn",
          "closed"
        ],
        "title": "ApplicationStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StatusUpdateRequest": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ApplicationStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "StatusUpdateRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "career-os",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/me/applications": {
      "get": {
        "operationId": "listApplications",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Application tracker",
        "tags": [
          "applications"
        ]
      }
    },
    "/v1/me/applications/{application_id}": {
      "patch": {
        "operationId": "updateApplicationStatus",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Application Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update application pipeline status",
        "tags": [
          "applications"
        ]
      }
    },
...
    "/v1/opportunities/{opportunity_id}/apply": {
      "post": {
        "operationId": "markOpportunityApplied",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm applied (user-confirmed; no auto-submit)",
        "tags": [
          "applications"
        ]
      }
    },
...
    "/v1/opportunities/{opportunity_id}/save": {
      "post": {
        "operationId": "saveOpportunity",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save / mark interested",
        "tags": [
          "applications"
        ]
      }
    }
  }
}